Add HelperMethods
